+++ pymongo/message.py
@@ -77,8 +77,8 @@ def _maybe_add_read_preference(spec, read_preference):
     # Only add $readPreference if it's something other than primary to avoid
     # problems with mongos versions that don't support read preferences. Also,
     # for maximum backwards compatibility, don't add $readPreference for
-    # secondaryPreferred unless tags or maxStalenessMS are in use (setting the
-    # slaveOkay bit has the same effect).
+    # secondaryPreferred unless tags or maxStalenessSeconds are in use (setting
+    # the slaveOkay bit has the same effect).
     if mode and (
         mode != ReadPreference.SECONDARY_PREFERRED.mode
         or tag_sets != [{}]