+++ doc/changelog.rst
@@ -13,7 +13,7 @@ Highlights include:
 
   - Unicode aware string comparison using :doc:`examples/collations`.
   - Support for the new :class:`~bson.decimal128.Decimal128` BSON type.
-  - A new maxStalenessMS read preference option.
+  - A new maxStalenessSeconds read preference option.
   - :meth:`~pymongo.collection.Collection.parallel_scan` supports maxTimeMS.
   - :attr:`~pymongo.write_concern.WriteConcern` is automatically
     applied by all helpers for commands that write to the database when